Fix HF checkpoint test to use the input test model
